refactor: move reused base serializer classes
We use a few convenience serializers across our MFE. Currently
CharListField, a list of strings, and IsRequiredField, a field
converting from text-based "required" string to boolean.

This moves them into a base utils file for reuse across our serializers.
